--- tests/transforms/test_optimization.py
@@ -1,876 +1,0 @@
-# Copyright 2018-2021 Xanadu Quantum Technologies Inc.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-#     http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import pytest
-import numpy as np
-
-import pennylane as qml
-from pennylane.wires import Wires
-from pennylane import apply
-
-from gate_data import I
-
-from pennylane.transforms.optimization import (
-    cancel_inverses,
-    commute_x_behind_targets,
-    commute_z_behind_controls,
-    merge_rotations,
-    single_qubit_fusion,
-)
-
-
-class TestCancelInverses:
-    """Test that adjacent inverse gates are cancelled."""
-
-    def test_one_qubit_cancel_adjacent_inverse(self):
-        """Test that a single-qubit circuit with adjacent gates cancels."""
-
-        def qfunc():
-            qml.Hadamard(wires=0)
-            qml.Hadamard(wires=0)
-
-        transformed_qfunc = cancel_inverses(qfunc)
-
-        new_tape = qml.transforms.make_tape(transformed_qfunc)()
-
-        assert len(new_tape.operations) == 0
-
-    def test_one_qubit_no_inverse(self):
-        """Test that a one-qubit circuit with a gate in the way does not cancel the inverses."""
-
-        def qfunc():
-            qml.Hadamard(wires=0)
-            qml.RZ(0.3, wires=0)
-            qml.Hadamard(wires=0)
-
-        transformed_qfunc = cancel_inverses(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 3
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires(0)
-
-        assert ops[1].name == "RZ"
-        assert ops[1].wires == Wires(0)
-
-        assert ops[2].name == "Hadamard"
-        assert ops[2].wires == Wires(0)
-
-    def test_two_qubits_no_inverse(self):
-        """Test that a two-qubit circuit self-inverse on each qubit does not cancel."""
-
-        def qfunc():
-            qml.Hadamard(wires=0)
-            qml.Hadamard(wires=1)
-
-        transformed_qfunc = cancel_inverses(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 2
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires(0)
-
-        assert ops[1].name == "Hadamard"
-        assert ops[1].wires == Wires(1)
-
-    def test_three_qubits_inverse_after_cnot(self):
-        """Test that a three-qubit circuit with a CNOT still allows cancellation."""
-
-        def qfunc():
-            qml.Hadamard(wires=0)
-            qml.PauliX(wires=1)
-            qml.CNOT(wires=[0, 2])
-            qml.RZ(0.5, wires=2)
-            qml.PauliX(wires=1)
-
-        transformed_qfunc = cancel_inverses(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 3
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires(0)
-
-        assert ops[1].name == "CNOT"
-        assert ops[1].wires == Wires([0, 2])
-
-        assert ops[2].name == "RZ"
-        assert ops[2].wires == Wires(2)
-
-    def test_three_qubits_blocking_cnot(self):
-        """Test that a three-qubit circuit with a blocking CNOT causes no cancellation."""
-
-        def qfunc():
-            qml.Hadamard(wires=0)
-            qml.PauliX(wires=1)
-            qml.CNOT(wires=[0, 1])
-            qml.RZ(0.5, wires=2)
-            qml.PauliX(wires=1)
-
-        transformed_qfunc = cancel_inverses(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 5
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires(0)
-
-        assert ops[1].name == "PauliX"
-        assert ops[1].wires == Wires(1)
-
-        assert ops[2].name == "CNOT"
-        assert ops[2].wires == Wires([0, 1])
-
-        assert ops[3].name == "RZ"
-        assert ops[3].wires == Wires(2)
-
-        assert ops[4].name == "PauliX"
-        assert ops[4].wires == Wires(1)
-
-    def test_two_qubits_cnot_same_direction(self):
-        """Test that two adjacent CNOTs cancel."""
-
-        def qfunc():
-            qml.CNOT(wires=[0, 1])
-            qml.CNOT(wires=[0, 1])
-
-        transformed_qfunc = cancel_inverses(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 0
-
-    def test_two_qubits_cnot_opposite_direction(self):
-        """Test that two adjacent CNOTs with the control/target flipped do NOT cancel."""
-
-        def qfunc():
-            qml.CNOT(wires=[0, 1])
-            qml.CNOT(wires=[1, 0])
-
-        transformed_qfunc = cancel_inverses(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 2
-
-        assert ops[0].name == "CNOT"
-        assert ops[0].wires == Wires([0, 1])
-
-        assert ops[1].name == "CNOT"
-        assert ops[1].wires == Wires([1, 0])
-
-    def test_two_qubits_cz_opposite_direction(self):
-        """Test that two adjacent CZ with the control/target flipped do cancel due to symmetry."""
-
-        def qfunc():
-            qml.CZ(wires=[0, 1])
-            qml.CZ(wires=[1, 0])
-
-        transformed_qfunc = cancel_inverses(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 0
-
-
-class TestCommuteXBehindTargets:
-    """Test that X rotations are properly pushed behind targets of X-based controlled operations."""
-
-    def test_single_x_after_cnot_gate(self):
-        """Test that a single X after a CNOT is pushed behind."""
-
-        def qfunc():
-            qml.Hadamard(wires=0)
-            qml.CNOT(wires=[0, 2])
-            qml.PauliX(wires=2)
-
-        transformed_qfunc = commute_x_behind_targets(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 3
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires(0)
-
-        assert ops[1].name == "PauliX"
-        assert ops[1].wires == Wires(2)
-
-        assert ops[2].name == "CNOT"
-        assert ops[2].wires == Wires([0, 2])
-
-    def test_multiple_x_after_cnot_gate(self):
-        """Test that multiple X rotations after a CNOT both get pushed behind."""
-
-        def qfunc():
-            qml.Hadamard(wires="a")
-            qml.CNOT(wires=["b", "a"])
-            qml.RX(0.2, wires="a")
-            qml.PauliX(wires="a")
-
-        transformed_qfunc = commute_x_behind_targets(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 4
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires("a")
-
-        assert ops[1].name == "RX"
-        assert ops[1].parameters[0] == 0.2
-        assert ops[1].wires == Wires("a")
-
-        assert ops[2].name == "PauliX"
-        assert ops[2].wires == Wires("a")
-
-        assert ops[3].name == "CNOT"
-        assert ops[3].wires == Wires(["b", "a"])
-
-    def test_single_x_after_crx_gate(self):
-        """Test that a single X rotation after a CRX is pushed behind."""
-
-        def qfunc():
-            qml.Hadamard(wires=0)
-            qml.CRX(0.1, wires=[0, "a"])
-            qml.RX(0.2, wires="a")
-
-        transformed_qfunc = commute_x_behind_targets(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 3
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires(0)
-
-        assert ops[1].name == "RX"
-        assert ops[1].parameters[0] == 0.2
-        assert ops[1].wires == Wires("a")
-
-        assert ops[2].name == "CRX"
-        assert ops[2].parameters[0] == 0.1
-        assert ops[2].wires == Wires([0, "a"])
-
-    def test_multiple_x_after_crx_gate(self):
-        """Test that multiple X rotations after a CRX are pushed behind."""
-
-        def qfunc():
-            qml.Hadamard(wires="a")
-            qml.CRX(0.3, wires=["b", "a"])
-            qml.PauliX(wires="a")
-            qml.RX(0.1, wires="a")
-
-        transformed_qfunc = commute_x_behind_targets(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 4
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires("a")
-
-        assert ops[1].name == "PauliX"
-        assert ops[1].wires == Wires("a")
-
-        assert ops[2].name == "RX"
-        assert ops[2].parameters[0] == 0.1
-        assert ops[2].wires == Wires("a")
-
-        assert ops[3].name == "CRX"
-        assert ops[3].parameters[0] == 0.3
-        assert ops[3].wires == Wires(["b", "a"])
-
-    def test_single_x_after_toffoli_gate(self):
-        """Test that a single X rotation after a Toffoli is pushed behind."""
-
-        def qfunc():
-            qml.Hadamard(wires=0)
-            qml.Toffoli(wires=[0, 3, "a"])
-            qml.RX(0.2, wires="a")
-
-        transformed_qfunc = commute_x_behind_targets(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 3
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires(0)
-
-        assert ops[1].name == "RX"
-        assert ops[1].parameters[0] == 0.2
-        assert ops[1].wires == Wires("a")
-
-        assert ops[2].name == "Toffoli"
-        assert ops[2].wires == Wires([0, 3, "a"])
-
-    def test_multiple_x_after_toffoli_gate(self):
-        """Test that multiple X rotations after a Toffoli are pushed behind."""
-
-        def qfunc():
-            qml.Hadamard(wires="a")
-            qml.Toffoli(wires=["b", "c", "a"])
-            qml.RX(0.1, wires="a")
-            qml.PauliX(wires="b")
-            qml.RX(0.2, wires="a")
-
-        transformed_qfunc = commute_x_behind_targets(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 5
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires("a")
-
-        assert ops[1].name == "RX"
-        assert ops[1].parameters[0] == 0.1
-        assert ops[1].wires == Wires("a")
-
-        assert ops[2].name == "RX"
-        assert ops[2].parameters[0] == 0.2
-        assert ops[2].wires == Wires("a")
-
-        assert ops[3].name == "Toffoli"
-        assert ops[3].wires == Wires(["b", "c", "a"])
-
-        assert ops[4].name == "PauliX"
-        assert ops[4].wires == Wires(["b"])
-
-    def test_no_commuting_gates_after_crx(self):
-        """Test that pushing commuting X gates behind targets is properly 'blocked'."""
-
-        def qfunc():
-            qml.Hadamard(wires=0)
-            qml.CRX(0.1, wires=[0, "a"])
-            # The Hadamard blocks the CRX from moving ahead of the PauliX
-            qml.Hadamard(wires="a")
-            qml.PauliX(wires="a")
-
-        transformed_qfunc = commute_x_behind_targets(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 4
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires(0)
-
-        assert ops[1].name == "CRX"
-        assert ops[1].parameters[0] == 0.1
-        assert ops[1].wires == Wires([0, "a"])
-
-        assert ops[2].name == "Hadamard"
-        assert ops[2].wires == Wires("a")
-
-        assert ops[3].name == "PauliX"
-        assert ops[3].wires == Wires("a")
-
-
-class TestCommuteZBehindControls:
-    """Test that diagonal gates are properly pushed behind X-based target operations."""
-
-    def test_single_z_after_cnot_gate(self):
-        """Test that a single Z after a CNOT is pushed behind."""
-
-        def qfunc():
-            qml.Hadamard(wires=0)
-            qml.CNOT(wires=[0, 2])
-            qml.PauliZ(wires=0)
-
-        transformed_qfunc = commute_z_behind_controls(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 3
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires(0)
-
-        assert ops[1].name == "PauliZ"
-        assert ops[1].wires == Wires(0)
-
-        assert ops[2].name == "CNOT"
-        assert ops[2].wires == Wires([0, 2])
-
-    def test_multiple_z_after_cnot_gate(self):
-        """Test that multiple Z rotations after a CNOT both get pushed behind."""
-
-        def qfunc():
-            qml.Hadamard(wires="a")
-            qml.CNOT(wires=["b", "a"])
-            qml.RZ(0.2, wires="b")
-            qml.PauliZ(wires="b")
-
-        transformed_qfunc = commute_z_behind_controls(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 4
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires("a")
-
-        assert ops[1].name == "RZ"
-        assert ops[1].parameters[0] == 0.2
-        assert ops[1].wires == Wires("b")
-
-        assert ops[2].name == "PauliZ"
-        assert ops[2].wires == Wires("b")
-
-        assert ops[3].name == "CNOT"
-        assert ops[3].wires == Wires(["b", "a"])
-
-    def test_single_z_after_cry_gate(self):
-        """Test that a single Z rotation after a CRY is pushed behind."""
-
-        def qfunc():
-            qml.Hadamard(wires=0)
-            qml.CRY(0.1, wires=[0, "a"])
-            qml.RZ(0.2, wires=0)
-
-        transformed_qfunc = commute_z_behind_controls(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 3
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires(0)
-
-        assert ops[1].name == "RZ"
-        assert ops[1].parameters[0] == 0.2
-        assert ops[1].wires == Wires(0)
-
-        assert ops[2].name == "CRY"
-        assert ops[2].parameters[0] == 0.1
-        assert ops[2].wires == Wires([0, "a"])
-
-    def test_multiple_z_after_crx_gate(self):
-        """Test that multiple Z rotations after a CRX are pushed behind."""
-
-        def qfunc():
-            qml.Hadamard(wires="a")
-            qml.CRX(0.3, wires=["b", "a"])
-            qml.PhaseShift(0.2, wires="b")
-            qml.T(wires="b")
-
-        transformed_qfunc = commute_z_behind_controls(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 4
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires("a")
-
-        assert ops[1].name == "PhaseShift"
-        assert ops[1].parameters[0] == 0.2
-        assert ops[1].wires == Wires("b")
-
-        assert ops[2].name == "T"
-        assert ops[2].wires == Wires("b")
-
-        assert ops[3].name == "CRX"
-        assert ops[3].parameters[0] == 0.3
-        assert ops[3].wires == Wires(["b", "a"])
-
-    def test_no_commuting_gates_after_crx(self):
-        """Test that pushing commuting X gates behind targets is properly 'blocked'."""
-
-        def qfunc():
-            qml.Hadamard(wires=0)
-            qml.CRX(0.1, wires=[0, "a"])
-            # The Hadamard blocks the CRX from moving ahead of the PauliX
-            qml.Hadamard(wires=0)
-            qml.S(wires=0)
-
-        transformed_qfunc = commute_z_behind_controls(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 4
-
-        assert ops[0].name == "Hadamard"
-        assert ops[0].wires == Wires(0)
-
-        assert ops[1].name == "CRX"
-        assert ops[1].parameters[0] == 0.1
-        assert ops[1].wires == Wires([0, "a"])
-
-        assert ops[2].name == "Hadamard"
-        assert ops[2].wires == Wires(0)
-
-        assert ops[3].name == "S"
-        assert ops[3].wires == Wires(0)
-
-
-class TestMergeRotations:
-    """Test that adjacent rotation gates of the same type will add the angles."""
-
-    @pytest.mark.parametrize(
-        ("theta_1", "theta_2", "expected_ops"),
-        [
-            (0.3, -0.2, [qml.RZ(0.1, wires=0)]),
-            (0.15, -0.15, []),
-        ],
-    )
-    def test_one_qubit_rotation_merge(self, theta_1, theta_2, expected_ops):
-        """Test that a single-qubit circuit with adjacent rotation along the same
-        axis either merge, or cancel if the angles sum to 0."""
-
-        def qfunc():
-            qml.RZ(theta_1, wires=0)
-            qml.RZ(theta_2, wires=0)
-
-        transformed_qfunc = merge_rotations(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == len(expected_ops)
-
-        # Check that all operations and parameter values are as expected
-        for op_obtained, op_expected in zip(ops, expected_ops):
-            assert op_obtained.name == op_expected.name
-            assert np.allclose(op_obtained.parameters, op_expected.parameters)
-
-    @pytest.mark.parametrize(
-        ("theta_1", "theta_2", "expected_ops"),
-        [
-            (0.3, -0.2, [qml.RZ(0.3, wires=0), qml.RZ(-0.2, wires=1)]),
-            (0.15, -0.15, [qml.RZ(0.15, wires=0), qml.RZ(-0.15, wires=1)]),
-        ],
-    )
-    def test_two_qubits_rotation_no_merge(self, theta_1, theta_2, expected_ops):
-        """Test that a two-qubit circuit with rotations on different qubits
-        do not get merged."""
-
-        def qfunc():
-            qml.RZ(theta_1, wires=0)
-            qml.RZ(theta_2, wires=1)
-
-        transformed_qfunc = merge_rotations(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == len(expected_ops)
-
-        for op_obtained, op_expected in zip(ops, expected_ops):
-            assert op_obtained.name == op_expected.name
-            assert np.allclose(op_obtained.parameters, op_expected.parameters)
-
-    @pytest.mark.parametrize(
-        ("theta_11", "theta_12", "theta_21", "theta_22", "expected_ops"),
-        [
-            (0.3, -0.2, 0.5, -0.8, [qml.RX(0.1, wires=0), qml.RY(-0.3, wires=1)]),
-            (0.3, -0.3, 0.7, -0.1, [qml.RY(0.6, wires=1)]),
-        ],
-    )
-    def test_two_qubits_rotation_no_merge(
-        self, theta_11, theta_12, theta_21, theta_22, expected_ops
-    ):
-        """Test that a two-qubit circuit with rotations on different qubits get merged."""
-
-        def qfunc():
-            qml.RX(theta_11, wires=0)
-            qml.RY(theta_21, wires=1)
-            qml.RX(theta_12, wires=0)
-            qml.RY(theta_22, wires=1)
-
-        transformed_qfunc = merge_rotations(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == len(expected_ops)
-
-        for op_obtained, op_expected in zip(ops, expected_ops):
-            assert op_obtained.name == op_expected.name
-            assert np.allclose(op_obtained.parameters, op_expected.parameters)
-
-    def test_one_qubits_rotation_blocked(self):
-        """Test that rotations on one-qubit separated by a "blocking" operation don't merge."""
-
-        def qfunc():
-            qml.RX(0.5, wires=0)
-            qml.Hadamard(wires=0)
-            qml.RX(0.4, wires=0)
-
-        transformed_qfunc = merge_rotations(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 3
-
-        assert ops[0].name == "RX"
-        assert ops[0].parameters[0] == 0.5
-
-        assert ops[1].name == "Hadamard"
-
-        assert ops[2].name == "RX"
-        assert ops[2].parameters[0] == 0.4
-
-    def test_two_qubits_rotation_blocked(self):
-        """Test that rotations on a two-qubit system separated by a "blocking" operation
-        don't merge."""
-
-        def qfunc():
-            qml.RX(-0.42, wires=0)
-            qml.CNOT(wires=[0, 1])
-            qml.RX(0.8, wires=0)
-
-        transformed_qfunc = merge_rotations(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 3
-
-        assert ops[0].name == "RX"
-        assert ops[0].wires == Wires(0)
-        assert ops[0].parameters[0] == -0.42
-
-        assert ops[1].name == "CNOT"
-        assert ops[1].wires == Wires([0, 1])
-
-        assert ops[2].name == "RX"
-        assert ops[2].wires == Wires(0)
-        assert ops[2].parameters[0] == 0.8
-
-    def test_controlled_rotation_merge(self):
-        """Test that adjacent controlled rotations on the same wires in same order get merged."""
-
-        def qfunc():
-            qml.CRY(0.2, wires=["w1", "w2"])
-            qml.CRY(0.3, wires=["w1", "w2"])
-
-        transformed_qfunc = merge_rotations(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 1
-
-        assert ops[0].name == "CRY"
-        assert ops[0].wires == Wires(["w1", "w2"])
-        assert ops[0].parameters[0] == 0.5
-
-    def test_controlled_rotation_no_merge(self):
-        """Test that adjacent controlled rotations on the same wires in different order don't merge."""
-
-        def qfunc():
-            qml.CRX(0.2, wires=["w1", "w2"])
-            qml.CRX(0.3, wires=["w2", "w1"])
-
-        transformed_qfunc = merge_rotations(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 2
-
-        assert ops[0].name == "CRX"
-        assert ops[0].wires == Wires(["w1", "w2"])
-        assert ops[0].parameters[0] == 0.2
-
-        assert ops[1].name == "CRX"
-        assert ops[1].wires == Wires(["w2", "w1"])
-        assert ops[1].parameters[0] == 0.3
-
-
-class TestSingleQubitFusion:
-    """Test fusion of groups of adjacent single-qubit gates."""
-
-    @pytest.mark.parametrize(
-        ("op_list"),
-        [
-            ([qml.Rot(0.0, 0.5, 0.0, wires=0), qml.RY(-0.5, wires=0)]),
-            ([qml.RZ(0.3, wires=3), qml.Rot(0.0, 0.0, -0.3, wires=3)]),
-            ([qml.Rot(0.0, 0.0, -0.3, wires=3), qml.RZ(0.3, wires=3)]),
-            ([qml.RZ(-np.pi / 2, wires=0), qml.S(wires=0)]),
-        ],
-    )
-    def test_single_qubit_fusion_zero_angles(self, op_list):
-        """Test a sequence of single-qubit operations that cancel upon fusion actually do so."""
-
-        def qfunc():
-            for op in op_list:
-                apply(op)
-
-        transformed_qfunc = single_qubit_fusion(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 0
-
-    @pytest.mark.parametrize(
-        ("op_list"),
-        [
-            ([qml.Rot(0.0, 0.0, 0.0, wires=3), qml.Rot(0.3, 0.4, 0.5, wires=3)]),
-            ([qml.Rot(0.0, 0.3, 0.0, wires=0), qml.RY(0.3, wires=0)]),
-            ([qml.RZ(0.1, wires=0), qml.RX(0.2, wires=0), qml.Hadamard(wires=0)]),
-            (
-                [
-                    qml.Rot(0.1, 0.2, 0.3, wires="a"),
-                    qml.PhaseShift(0.2, wires="a"),
-                    qml.S(wires="a"),
-                    qml.RY(0.2, wires="a"),
-                ]
-            ),
-        ],
-    )
-    def test_single_qubit_fusion_same_wire(self, op_list):
-        """Test a sequence of single-qubit operations on the same wire are fused to a single Rot"""
-
-        def qfunc():
-            for op in op_list:
-                apply(op)
-
-        transformed_qfunc = single_qubit_fusion(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 1
-        assert ops[0].name == "Rot"
-
-        # Compare matrix representations (up to a global phase)
-        expected_mat = I
-        for op in op_list:
-            expected_mat = np.dot(op.matrix, expected_mat)
-
-        obtained_mat = ops[0].matrix
-
-        # Check equivalence by seeing if U^\dagger U is I up to a phase
-        mat_product = np.dot(np.conj(obtained_mat.T), expected_mat)
-        mat_product /= mat_product[0, 0]
-
-        assert qml.math.allclose(mat_product, I)
-
-    @pytest.mark.parametrize(
-        ("op_list_q1,op_list_q2"),
-        [
-            (
-                [qml.Rot(0.0, 0.0, 0.0, wires=3), qml.Rot(0.3, 0.4, 0.5, wires=3)],
-                [qml.RZ(0.2, wires=2), qml.T(wires=2), qml.PauliX(wires=2)],
-            ),
-            (
-                [qml.PauliX(wires="a"), qml.PauliZ(wires="a"), qml.RY(-0.2, wires="a")],
-                [qml.RZ(0.2, wires="b"), qml.S(wires="b"), qml.PhaseShift(0.5, wires="b")],
-            ),
-            (
-                [qml.PauliY(wires=0), qml.PauliZ(wires=0), qml.PauliY(wires=0)],
-                [qml.SX(wires="b"), qml.RX(0.0, wires="b")],
-            ),
-        ],
-    )
-    def test_single_qubit_fusion_two_wires(self, op_list_q1, op_list_q2):
-        """Test that independent sequences of rotations on different qubits properly fuse."""
-
-        def qfunc():
-            for op in op_list_q1 + op_list_q2:
-                apply(op)
-
-        transformed_qfunc = single_qubit_fusion(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 2
-
-        assert ops[0].name == "Rot"
-        assert ops[1].name == "Rot"
-
-        # Compare matrix representations (up to a global phase)
-        expected_mat_q1 = I
-        for op in op_list_q1:
-            expected_mat_q1 = np.dot(op.matrix, expected_mat_q1)
-
-        expected_mat_q2 = I
-        for op in op_list_q2:
-            expected_mat_q2 = np.dot(op.matrix, expected_mat_q2)
-
-        # Check equivalence by seeing if U^\dagger U is I up to a phase
-        mat_product_1 = np.dot(np.conj(ops[0].matrix.T), expected_mat_q1)
-        mat_product_1 /= mat_product_1[0, 0]
-
-        mat_product_2 = np.dot(np.conj(ops[1].matrix.T), expected_mat_q2)
-        mat_product_2 /= mat_product_2[0, 0]
-
-        assert qml.math.allclose(mat_product_1, I)
-        assert qml.math.allclose(mat_product_2, I)
-
-    @pytest.mark.parametrize(
-        ("op_list_1,op_list_2"),
-        [
-            (
-                [qml.Rot(0.0, 0.0, 0.0, wires=3), qml.Rot(0.3, 0.4, 0.5, wires=3)],
-                [qml.RZ(0.2, wires=3), qml.T(wires=3), qml.PauliX(wires=3)],
-            ),
-            (
-                [qml.PauliX(wires="a"), qml.PauliZ(wires="a"), qml.RY(-0.2, wires="a")],
-                [qml.RZ(0.2, wires="a"), qml.S(wires="a"), qml.PhaseShift(0.5, wires="a")],
-            ),
-            (
-                [qml.PauliY(wires=0), qml.PauliZ(wires=0), qml.PauliY(wires=0)],
-                [qml.SX(wires=0), qml.RX(0.0, wires=0)],
-            ),
-        ],
-    )
-    def test_single_qubit_fusion_blocked_by_cnot(self, op_list_1, op_list_2):
-        """Test that sequences of rotations on each side of a CNOT fuse independently."""
-
-        def qfunc():
-            for op in op_list_1:
-                apply(op)
-
-            qml.CNOT(wires=[op_list_1[0].wires[0], "c"])
-
-            for op in op_list_2:
-                apply(op)
-
-        transformed_qfunc = single_qubit_fusion(qfunc)
-
-        ops = qml.transforms.make_tape(transformed_qfunc)().operations
-
-        assert len(ops) == 3
-
-        assert ops[0].name == "Rot"
-        assert ops[0].wires[0] == op_list_1[0].wires[0]
-
-        assert ops[1].name == "CNOT"
-        assert ops[1].wires == Wires([op_list_1[0].wires[0], "c"])
-
-        assert ops[2].name == "Rot"
-        assert ops[2].wires[0] == op_list_2[0].wires[0]
-
-        # Compare matrix representations (up to a global phase)
-        expected_mat_1 = I
-        for op in op_list_1:
-            expected_mat_1 = np.dot(op.matrix, expected_mat_1)
-
-        expected_mat_2 = I
-        for op in op_list_2:
-            expected_mat_2 = np.dot(op.matrix, expected_mat_2)
-
-        # Check equivalence by seeing if U^\dagger U is I up to a phase
-        mat_product_1 = np.dot(np.conj(ops[0].matrix.T), expected_mat_1)
-        mat_product_1 /= mat_product_1[0, 0]
-
-        mat_product_2 = np.dot(np.conj(ops[2].matrix.T), expected_mat_2)
-        mat_product_2 /= mat_product_2[0, 0]
-
-        assert qml.math.allclose(mat_product_1, I)
-        assert qml.math.allclose(mat_product_2, I)
